separate out HsTemplate to HsTemplate and HsTH.
